requirements.txt
whisperx
pyannote.audio
torch
numpy
scipy
